win-dshow: Add C985/C353 hardware h264 encoders
This adds support for the AverMedia C985 encoder (which is available on
C985 capture cards) as well as the C353 hardware encoder (which is
currently available on the X99S Gaming 9 motherboards).

These encoders have some limitations, such as a max resolution of
1280x720, a max GOP size of 30, and the encoder format only supports
YV12, which requires conversion if the current output format isn't the
same.  The C985 and C353 encoders seem to be pretty much identical,
although it seems like the C353 has a bit more efficient encoding.

I don't believe these are really suitable for streaming, as they do not
really have the encoding efficiency needed to stream at lower bitrates,
and seem to only support variable bitrate.  However, for recording these
encoders are quite nice to have available, and work quite well.